backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.2.0
beanie==1.26.0
click==8.1.7
dnspython==2.6.1
fastapi==0.114.2
h11==0.14.0
idna==3.9
lazy-model==0.2.0
motor==3.5.1
passlib==1.7.4
pydantic==2.9.1
pydantic_core==2.23.3
pymongo==4.8.0
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.38.5
toml==0.10.2
typing_extensions==4.12.2
uvicorn==0.30.6
